Änderung der gesamten Ordnerstruktur/Dynamische Lins/Rollenverteilungen etc.
