+++ app/[instanz]/[organisation]/[user]/page.js
@@ -1,0 +1,13 @@
+"use client";
+
+
+import React, { useState } from 'react';
+import { Box } from '@mui/material';
+
+export default function AdminDashboard() {
+  return (
+    <Box sx={{ display: 'flex', height: '100vh' }}>
+        Hallo, das ist eine Testseite!
+    </Box>
+  );
+}
