--- app/admin/page.js
@@ -1,13 +1,0 @@
-"use client";
-
-
-import React, { useState } from 'react';
-import { Box } from '@mui/material';
-
-export default function AdminDashboard() {
-  return (
-    <Box sx={{ display: 'flex', height: '100vh' }}>
-        Hallo, ich bin die Admin-Seite
-    </Box>
-  );
-}
