+++ app/components/Sidebar.js
@@ -1,14 +1,76 @@
 "use client";
 
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
-import PropTypes from "prop-types";
 import SidebarDesign from "../components/styledComponents/SidebarDesign";
+import { useUserContext } from "../context/UserContext"; // Benutzerkontext
 
-function Sidebar({ role }) {
+function Sidebar() {
   const [isExpanded, setIsExpanded] = useState(false);
   const [drawerOpen, setDrawerOpen] = useState(false); // Zustand für das mobile Menü
   const router = useRouter();
+  const { userInfo } = useUserContext(); // Benutzerinformationen aus dem Kontext
+  const [role, setRole] = useState("");
+
+  // Rolle aus dem Benutzerkontext setzen
+  useEffect(() => {
+    if (userInfo && userInfo.role) {
+      setRole(userInfo.role);
+    }
+  }, [userInfo]);
+
+  if (!userInfo) {
+    return null; // Zeige nichts, wenn der Benutzer nicht eingeloggt ist
+  }
+
+  const basePath = `/${userInfo.instanz}/${userInfo.organisation}/${userInfo.username}`;
+
+  let mainMenuItems = [];
+  let bottomMenuItems = [
+    { icon: "Settings", text: "Einstellungen", action: () => router.push(`${basePath}/settings`) }
+  ];
+
+  let dashboard = { icon: "Dashboard", text: "Dashboard", action: () => router.push(basePath) }
+  switch (role) {
+    case "organisation-admin":
+      mainMenuItems = [
+        dashboard,
+        { icon: "Group", text: "Benutzerverwaltung", action: () => router.push(`${basePath}/userControl`) },
+        { icon: "Event", text: "Terminmanagement", action: () => router.push(`${basePath}/termin`) },
+        { icon: "Notifications", text: "Benachrichtigungen", action: () => router.push(`${basePath}/notification`) },
+        { icon: "ReportProblem", text: "Ankündigungen", action: () => router.push(`${basePath}/announcements`) },
+        { icon: "Group", text: "Veranstaltung erstellen", action: () => router.push(`${basePath}/createEvent`) },
+        { icon: "Event", text: "Meine Veranstaltungen", action: () => router.push(`${basePath}/myevent`) },
+        { icon: "Event", text: "Meine Teilnahmen", action: () => router.push(`${basePath}/myParticipations`) },
+        { icon: "Bookmark", text: "Mein Entwurf", action: () => router.push(`${basePath}/meinEntwurf`) },
+      ];
+      break;
+
+    case "organisator":
+      mainMenuItems = [
+       dashboard,
+        { icon: "Event", text: "Terminmanagement", action: () => router.push(`${basePath}/termin`) },
+        { icon: "Notifications", text: "Benachrichtigungen", action: () => router.push(`${basePath}/notification`) },
+        { icon: "ReportProblem", text: "Ankündigungen", action: () => router.push(`${basePath}/announcements`) },
+        { icon: "Group", text: "Veranstaltung erstellen", action: () => router.push(`${basePath}/createEvent`) },
+        { icon: "Event", text: "Meine Veranstaltungen", action: () => router.push(`${basePath}/myevent`) },
+        { icon: "Event", text: "Meine Teilnahmen", action: () => router.push(`${basePath}/myParticipations`) },
+        { icon: "Bookmark", text: "Mein Entwurf", action: () => router.push(`${basePath}/meinEntwurf`) },
+      ];
+      break;
+
+      case "teilnehmer":
+        mainMenuItems = [
+          dashboard,
+          { icon: "Event", text: "Terminmanagement", action: () => router.push(`${basePath}/termin`) },
+          { icon: "Event", text: "Meine Veranstaltungen", action: () => router.push(`${basePath}/myevent`) },
+        ];
+        break;
+        
+
+    default:
+      mainMenuItems = []; // Optional: Menüeinträge für unbekannte Rollen
+  }
 
   const handleMouseEnter = () => {
     setIsExpanded(true);
@@ -22,32 +84,6 @@ function Sidebar({ role }) {
     setDrawerOpen(!drawerOpen);
   };
 
-  let mainMenuItems = [];
-
-  switch (role) {
-    case "admin":
-      mainMenuItems = [
-        { icon: "Dashboard", text: "Dashboard", action: () => router.push("") },
-        { icon: "Group", text: "Benutzerverwaltung", action: () => router.push("/admin/users") },
-        { icon: "Event", text: "Terminmanagement", action: () => router.push("/admin/termin") },
-        { icon: "Notifications", text: "Benachrichtigungen", action: () => router.push("/admin/notification") },
-        { icon: "ReportProblem", text: "Ankündigungen", action: () => router.push("/admin/announcements") },
-      ];
-      break;
-
-    case "user":
-      mainMenuItems = [
-        { icon: "Group", text: "Veranstaltung erstellen", action: () => router.push("/user/createEvent") },
-        { icon: "Event", text: "Meine Veranstaltungen", action: () => router.push("/user/myevent") },
-        { icon: "Event", text: "Meine Teilnahmen", action: () => {} },
-        { icon: "Bookmark", text: "Mein Entwurf", action: () => router.push("/user/meinEntwurf") },
-      ];
-      break;
-
-    default:
-      mainMenuItems = []; // Optional: Menüeinträge für unbekannte Rollen oder keine Rolle
-  }
-
   return (
     <SidebarDesign
       isExpanded={isExpanded}
@@ -56,13 +92,10 @@ function Sidebar({ role }) {
       drawerOpen={drawerOpen}
       toggleDrawer={toggleDrawer}
       mainMenuItems={mainMenuItems}
+      bottomMenuItems={bottomMenuItems} // Übergabe des Bottom-Menüs
       role={role}
     />
   );
 }
-
-Sidebar.propTypes = {
-  role: PropTypes.oneOf(["admin", "user", "participant"]).isRequired,
-};
 
 export default Sidebar;
