+++ app/layout.js
@@ -5,7 +5,7 @@ import { UserProvider } from '@auth0/nextjs-auth0/client';
 import {Box} from "@mui/material";
 import { ChatProvider } from "./components/ChatContext";
 import { DarkModeProvider } from "./components/styledComponents/DarkMode"
-
+import {UserProviderr} from "./context/UserContext"
 
 export const metadata = {
   title: "Veranstaltungsplaner",
@@ -17,10 +17,12 @@ export default function RootLayout({ children }) {
     <html lang="en">
       <DarkModeProvider>
       <body>
-        <UserProvider>
+      <UserProvider>
+        <UserProviderr>
           <ChatProvider>
             {children}
           </ChatProvider>
+        </UserProviderr>
         </UserProvider>
       </body>
       </DarkModeProvider>
