+++ app/page.js
@@ -1,5 +1,6 @@
 import LogInOut from '../app/components/LogInOut';
+import SimpleLogin from "./components/SimpleLogin"
 
 export default function HomePage() {
-  return <LogInOut />;
+  return <SimpleLogin />;
 }
